requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.23.2
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.1.0

# Image processing
opencv-python==4.8.1.78
Pillow==10.0.1
scipy==1.11.3
numpy==1.24.3
dlib==19.24.2

# HTTP and utilities
requests==2.31.0
python-dotenv==1.0.0
aiofiles==23.2.1

# Data handling
pandas==2.0.3